requirements.txt
openai
tiktoken
python-dotenv
chromadb
tqdm